# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Folder Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /folders:
    get:
      operationId: listFolders
      tags: [folders]
      summary: List folders
      parameters:
        - name: parentId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: flat
          in: query
          required: false
          schema:
            type: boolean
          description: If true, return all folders regardless of parent
      responses:
        "200":
          description: List of folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Folder"
    post:
      operationId: createFolder
      tags: [folders]
      summary: Create a new folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderInput"
      responses:
        "201":
          description: Created folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /folders/reorder:
    post:
      operationId: reorderFolders
      tags: [folders]
      summary: Reorder folders by setting positions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderReorderInput"
      responses:
        "200":
          description: Reordered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
  /folders/{id}:
    get:
      operationId: getFolder
      tags: [folders]
      summary: Get a single folder by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateFolder
      tags: [folders]
      summary: Update a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderUpdate"
      responses:
        "200":
          description: Updated folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteFolder
      tags: [folders]
      summary: Delete a folder and all its descendants
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /folders/{id}/breadcrumb:
    get:
      operationId: getFolderBreadcrumb
      tags: [folders]
      summary: Get breadcrumb path for a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Breadcrumb path
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FolderBreadcrumb"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Folder:
      type: object
      required:
        - id
        - name
        - color
        - icon
        - style
        - position
        - childCount
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        parentId:
          type: ["integer", "null"]
        color:
          type: string
        icon:
          type: string
          description: Lucide icon name (e.g. briefcase, star, heart)
        style:
          type: string
          description: "Card style: default | square | big | flat | wide"
        position:
          type: integer
          description: Sort order position (lower = first)
        childCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    FolderBreadcrumb:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
    FolderInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        parentId:
          type: ["integer", "null"]
        color:
          type: string
        icon:
          type: string
        style:
          type: string
    FolderUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        icon:
          type: string
        style:
          type: string
        parentId:
          type: ["integer", "null"]
    FolderReorderInput:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            type: object
            required:
              - id
              - position
            properties:
              id:
                type: integer
              position:
                type: integer
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string